API: Add file.getBlob() method
Get a W3C `Blob` object which contains the file data.

The file will be fetched from the network with highest priority, and
`callback` will be
called once the file is ready. `callback` must be specified, and will
be called with a an
`Error` (or `null`) and the `Blob` object.